+++ Source/DynamicScene/DataSourceDisplay.js
@@ -83,7 +83,7 @@ define(['../Core/defaultValue',
         this._staticSourcesToUpdate = [];
         this._visualizersCallback = defaultValue(visualizersCallback, createDefaultVisualizers);
 
-        for (var i = 0, len = dataSourceCollection.getLength(); i < len; i++) {
+        for (var i = 0, len = dataSourceCollection.length; i < len; i++) {
             this._onDataSourceAdded(dataSourceCollection, dataSourceCollection.get(i));
         }
     };
@@ -139,7 +139,7 @@ define(['../Core/defaultValue',
         this._eventHelper.removeAll();
 
         var dataSourceCollection = this._dataSourceCollection;
-        for (var i = 0, length = dataSourceCollection.getLength(); i < length; ++i) {
+        for (var i = 0, length = dataSourceCollection.length; i < length; ++i) {
             this._onDataSourceRemoved(this._dataSourceCollection, dataSourceCollection.get(i));
         }