+++ Specs/Core/FeatureDetectionSpec.js
@@ -6,11 +6,6 @@ defineSuite([
     "use strict";
     /*global jasmine,describe,xdescribe,it,xit,expect,beforeEach,afterEach,beforeAll,afterAll,spyOn,runs,waits,waitsFor*/
 
-    it('detects cross-origin imagery', function() {
-        //just make sure the function runs, the test can't expect a value of true or false
-        expect(FeatureDetection.supportsCrossOriginImagery()).toBeDefined();
-    });
-
     it('detects fullscreen support', function() {
         //just make sure the function runs, the test can't expect a value of true or false
         expect(FeatureDetection.supportsFullscreen()).toBeDefined();