+++ CHANGES.md
@@ -4,6 +4,21 @@ Change Log
 Beta Releases
 -------------
 
+### b27 - 2014-04-01
+
+* Breaking changes:
+  * All `CameraController` functions have been moved up to the `Camera`. Removed `CameraController`. For example, code that looked like:
+
+           scene.camera.controller.viewExtent(extent);
+
+    should now look like:
+
+           scene.camera.viewExtent(extent);
+
+* `loadArrayBuffer`, `loadBlob`, `loadJson`, `loadText`, and `loadXML` now support loading data from data URIs.
+* Fixed extruded polygons rendered in the southern hemisphere. [#1490](https://github.com/AnalyticalGraphicsInc/cesium/issues/1490)
+* Fixed Primitive picking that have a closed appearance drawn on the surface. [#1333](https://github.com/AnalyticalGraphicsInc/cesium/issues/1333)
+
 ### b26 - 2014-03-03
 
 * Breaking changes:
@@ -29,7 +44,7 @@ Beta Releases
       * `getMaximumRadius` -> `maximumRadius`
     * `CentralBody`
       * `getEllipsoid` -> `ellipsoid`
-      * `getImageryLayers` -> `imageryLayerCollection`
+      * `getImageryLayers` -> `imageryLayers`
     * `EllipsoidalOccluder`
       * `getEllipsoid` -> `ellipsoid`
       * `getCameraPosition`, `setCameraPosition` -> `cameraPosition`
@@ -82,6 +97,18 @@ Beta Releases
       * `getCredit` -> `credit`
       * `getTilingScheme` -> `tilingScheme`
       * `isReady` -> `ready`
+    * `TimeIntervalCollection`
+      * `getChangedEvent` -> `changedEvent`
+      * `getStart` -> `start`
+      * `getStop` -> `stop`
+      * `getLength` -> `length`
+      * `isEmpty` -> `empty`
+    * `DataSourceCollection`, `ImageryLayerCollection`, `LabelCollection`, `PolylineCollection`, `SensorVolumeCollection`
+      * `getLength` -> `length`
+    * `BillboardCollection`
+      * `getLength` -> `length`
+      * `getTextureAtlas`, `setTextureAtlas` -> `textureAtlas`
+      * `getDestroyTextureAtlas`, `setDestroyTextureAtlas` -> `destroyTextureAtlas`
   * Removed `Scene.getUniformState()`.  Use `scene.context.getUniformState()`.
   * Visualizers no longer create a `dynamicObject` property on the primitives they create.  Instead, they set the `id` property that is standard for all primitives.
   * The `propertyChanged` on DynamicScene objects has been renamed to `definitionChanged`.  Also, the event is now raised in the case of an existing property being modified as well as having a new property assigned (previously only property assignment would raise the event).
@@ -92,6 +119,7 @@ Beta Releases
   * Modified `CameraFlightPath` functions to take place in the camera's current reference frame. The arguments to the function now need to be given in world coordinates and an optional reference frame can be given when the flight is completed.
   * `PixelDatatype` properties are now JavaScript numbers, not `Enumeration` instances.
   * `combine` now takes two objects instead of an array, and defaults to copying shallow references.  The `allowDuplicates` parameter has been removed.  In the event of duplicate properties, the first object's properties will be used.
+  * Removed `FeatureDetection.supportsCrossOriginImagery`.  This check was only useful for very old versions of WebKit.
 * Added `Model` for drawing 3D models using glTF.  See the [tutorial](http://cesiumjs.org/2014/03/03/Cesium-3D-Models-Tutorial/) and [Sandcastle example](http://cesiumjs.org/Cesium/Apps/Sandcastle/index.html?src=3D%20Models.html&label=Showcases).
 * DynamicScene now makes use of [Geometry and Appearances](http://cesiumjs.org/2013/11/04/Geometry-and-Appearances/), which provides a tremendous improvements to DataSource visualization (CZML, GeoJSON, etc..).  Extruded geometries are now supported and in many use cases performance is an order of magnitude faster.
 * Added new `SelectionIndicator` and `InfoBox` widgets to `Viewer`, activated by `viewerDynamicObjectMixin`.